Added a 'contact-us' section to index page
I used Google maps embed and forms to make a contact-us section on the home page for customers to arrange booking, enquiries and more.

I made sure to keep the professional look of the homepage by including a large top margin on the container and by styling it in conjunction with the rest of the site.