Added a 'How To Order' button
I've added the basis for a How To Order button on the Order page which, when pressed, will display a clear instruction set for how to order if a customer is stuck how to.

I have done this in order to make the customer experience more intuitive and straightforward. while also relieving the waiters having to explain the online ordering system.