Added instructions in the how to order popup
I replaced the 'test' text with a clear., written instruction set for the customer to be able to read if they need to know how our ordering system works.

I plan to develop on this further in the future by simplifying the look of the prompt via the use of illustrations to make it easier for the customer to understand.